Move relative time helper to TimeUtils
