+++ app/src/main/java/org/ole/planet/myplanet/utilities/TimeUtils.kt
@@ -1,5 +1,6 @@
 package org.ole.planet.myplanet.utilities
 
+import android.text.format.DateUtils
 import java.time.Instant
 import java.time.LocalDate
 import java.time.LocalDateTime
@@ -155,4 +156,11 @@ object TimeUtils {
             e.printStackTrace()
             null
         }
+
+    fun getRelativeTime(timestamp: Long): String {
+        val timeNow = System.currentTimeMillis()
+        return if (timestamp < timeNow) {
+            DateUtils.getRelativeTimeSpanString(timestamp, timeNow, 0).toString()
+        } else "Just now"
+    }
 }
